+++ coding/ttr_map_maker/task_export_gui.py
@@ -22,7 +22,7 @@ class Task_Export_GUI:
   def __init__(self,
       master: tk.Tk,
       color_config: dict[str, str],
-      grid_padding: Tuple[float, float],
+      grid_padding: Tuple[int, int],
       tk_config_methods: dict[str, Callable],
       particle_graph: TTR_Particle_Graph,
       task_export_frame: tk.Frame,
